Extracted Email interface
