+++ src/main/java/com/itquasar/multiverse/mail/message/ImmutableEmail.java
@@ -1,9 +1,8 @@
 package com.itquasar.multiverse.mail.message;
 
-import com.itquasar.multiverse.mail.Content;
+import com.itquasar.multiverse.mail.message.content.Content;
 import com.itquasar.multiverse.mail.EmailContact;
-import com.itquasar.multiverse.mail.Envelope;
-import com.itquasar.multiverse.mail.NewInterface;
+import com.itquasar.multiverse.mail.message.envelope.Envelope;
 import com.itquasar.multiverse.mail.message.content.LazyContent;
 import com.itquasar.multiverse.mail.message.envelope.ImmutableEnvelope;
 import com.itquasar.multiverse.mail.message.envelope.LazyEnvelope;
@@ -29,7 +28,7 @@ import org.slf4j.LoggerFactory;
  *
  * @author Guilherme I F L Weizenmann <guilherme at itquasar.com>
  */
-public class ImmutableEmail implements NewInterface {
+public class ImmutableEmail implements Email {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ImmutableEmail.class);
