+++ src/main/java/com/itquasar/multiverse/mail/message/content/ImmutableContent.java
@@ -1,6 +1,5 @@
 package com.itquasar.multiverse.mail.message.content;
 
-import com.itquasar.multiverse.mail.Content;
 import com.itquasar.multiverse.mail.part.MimeTypes;
 import com.itquasar.multiverse.mail.part.Part;
 import com.itquasar.multiverse.mail.part.SinglePart;
