+++ src/main/java/com/itquasar/multiverse/mail/message/envelope/Envelope.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.itquasar.multiverse.mail;
+package com.itquasar.multiverse.mail.message.envelope;
 
+import com.itquasar.multiverse.mail.EmailContact;
 import java.util.Date;
 import java.util.List;
