+++ src/main/java/com/itquasar/multiverse/mail/message/envelope/ImmutableEnvelope.java
@@ -1,7 +1,6 @@
 package com.itquasar.multiverse.mail.message.envelope;
 
 import com.itquasar.multiverse.mail.EmailContact;
-import com.itquasar.multiverse.mail.Envelope;
 import com.itquasar.multiverse.mail.util.ClientUtils;
 import com.itquasar.multiverse.mail.util.FunctionUtils;
 import java.util.Collections;
